pyproject.toml
[project]
name = "energy-rte-tempo-forecasting"
version = "0.1.0"
description = "Energy forecasting — Projet 705"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
]

[project.optional-dependencies]
ci = [
    "pytest",
    "pyyaml",
    "fastapi",
    "httpx",
    "pydantic",
    "requests",
    "yamllint",
]
dev = [
    "pytest",
    "pyyaml",
    "pre-commit",
    "fastapi",
    "httpx",
    "pydantic",
    "requests",
]
